+++ src/main/java/school/faang/user_service/validator/mentorship/CheckingExistUserValidator.java
@@ -1,15 +1,17 @@
-package school.faang.user_service.validator;
+package school.faang.user_service.validator.mentorship;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ResponseStatusException;
 import school.faang.user_service.dto.mentorship.MentorshipRequestDto;
 import school.faang.user_service.repository.mentorship.MentorshipRepository;
+import school.faang.user_service.validator.Validator;
+
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Component
 @RequiredArgsConstructor
-public class CheckingExistUserValidator implements MentorshipValidator<MentorshipRequestDto> {
+public class CheckingExistUserValidator implements Validator<MentorshipRequestDto> {
 
     public final MentorshipRepository mentorshipRepository;
